+++ chaincode/lib/keyContract.js
@@ -11,14 +11,14 @@ const Key = require("./ledger/key");
 const KeyList = require("./ledger/keyList");
 
 
-class KeyContext extends Context{
+class KeyContext extends Context {
   constructor() {
     super();
     this.keyList = new KeyList(this);
   }
 }
 
-class KeyContract extends Contract{
+class KeyContract extends Contract {
 
   constructor() {
     super("org.eer.key");
@@ -39,30 +39,43 @@ class KeyContract extends Contract{
     console.info("=============  END : Initialize Ledger  ===========");
   }
 
-  async addKey(ctx,id,key){
-    let _key = Key.createInstance(id,key);
+  async addKey(ctx, id, key) {
+
+    if (await ctx.keyList.getKey(makeKeyKey(id))) {
+      return false;
+    }
+
+    let _key = Key.createInstance(id, key);
 
     await ctx.keyList.addKey(_key);
 
     return _key;
   }
 
-  async getKey(ctx,id){
+  async getKey(ctx, id) {
     let keyKey = makeKeyKey(id);
 
-    return await ctx.keyList.getKey(keyKey);
+    let _key = await ctx.keyList.getKey(keyKey);
+
+    if (_key) return _key.getKey();
+    else return null;
   }
 
-  async deleteKey(ctx, id){
+  async deleteKey(ctx, id) {
     let keyKey = makeKeyKey(id);
 
-    await ctx.keyList.deleteKey(keyKey);
+    let _key = ctx.keyList.getKey(keyKey);
+
+    if (_key) {
+      await ctx.keyList.deleteKey(keyKey);
+      return _key.getKey();
+    } else return false;
   }
 
 }
 
-function makeKeyKey(id){
-  return Key.makeKey(["KEY",id]);
+function makeKeyKey(id) {
+  return Key.makeKey(["KEY", id]);
 }
 
-module.exports= KeyContract;
+module.exports = KeyContract;
